Renamed "controllers.py" to "devices.py". (#97)
